Moved CPU & GPU settings into separate Tuning window
