+++ src/I18n/Languages/ChineseSimplified.cs
@@ -58,7 +58,10 @@ public static class ChineseSimplified
         // KEYBOARD
         ["keyboard_header"] = "笔记本键盘",
         ["backlight"] = "背光",
-        ["extra"] = "更多",
+        ["tuning_button"] = "Tuning",
+        ["tuning_header"] = "Tuning",
+        ["cpu_tab"] = "CPU",
+        ["gpu_tab"] = "GPU",
         ["backlight_off"] = "关",
         ["backlight_low"] = "低",
         ["backlight_medium"] = "中",
