+++ src/I18n/Languages/Danish.cs
@@ -58,7 +58,10 @@ public static class Danish
         // KEYBOARD
         ["keyboard_header"] = "Bærbart tastatur",
         ["backlight"] = "Baggrundsbelysning",
-        ["extra"] = "Ekstra",
+        ["tuning_button"] = "Tuning",
+        ["tuning_header"] = "Tuning",
+        ["cpu_tab"] = "CPU",
+        ["gpu_tab"] = "GPU",
         ["backlight_off"] = "Fra",
         ["backlight_low"] = "Lav",
         ["backlight_medium"] = "Medium",
