+++ src/UI/Views/ExtraWindow.axaml
@@ -190,51 +190,6 @@
             </Border>
 
 
-            <Border Classes="panel" x:Name="panelGpuTuning" IsVisible="False">
-                <StackPanel>
-                    <Border Classes="panel-header" Margin="0,0,0,8">
-                        <TextBlock x:Name="headerGpuTuning" Classes="header" Text="GPU Tuning" />
-                    </Border>
-
-                    <TextBlock x:Name="labelGpuTuningInfo" Classes="label-dim"
-                               Text="NVIDIA GPU" Margin="0,0,0,6" />
-
-                    <!-- Power Limit slider -->
-                    <Grid ColumnDefinitions="110,*,50" Margin="0,4">
-                        <TextBlock Grid.Column="0" x:Name="labelPowerLimitLabel" Text="Power Limit"
-                                   Classes="label-dim" VerticalAlignment="Center" />
-                        <Slider Grid.Column="1" x:Name="sliderGpuPowerLimit"
-                                Minimum="5" Maximum="115" Value="55"
-                                TickFrequency="5" IsSnapToTickEnabled="True"
-                                ValueChanged="SliderGpuPowerLimit_ValueChanged" />
-                        <TextBlock Grid.Column="2" x:Name="labelGpuPowerLimit"
-                                   Classes="value" Text="55W"
-                                   HorizontalAlignment="Center" VerticalAlignment="Center" />
-                    </Grid>
-
-                    <!-- Clock Lock -->
-                    <Grid ColumnDefinitions="110,Auto,*,50" Margin="0,4">
-                        <TextBlock Grid.Column="0" x:Name="labelClockLockLabel" Text="Clock Lock"
-                                   Classes="label-dim" VerticalAlignment="Center" />
-                        <CheckBox Grid.Column="1" x:Name="checkGpuClockLock"
-                                  Margin="0,0,8,0" VerticalAlignment="Center"
-                                  IsCheckedChanged="CheckGpuClockLock_Changed" />
-                        <Slider Grid.Column="2" x:Name="sliderGpuClockLock"
-                                Minimum="200" Maximum="3000" Value="2400"
-                                TickFrequency="50" IsSnapToTickEnabled="True"
-                                IsEnabled="False"
-                                ValueChanged="SliderGpuClockLock_ValueChanged" />
-                        <TextBlock Grid.Column="3" x:Name="labelGpuClockLock"
-                                   Classes="value" Text="--"
-                                   HorizontalAlignment="Center" VerticalAlignment="Center" />
-                    </Grid>
-
-                    <Button x:Name="buttonGpuApply" Classes="ghelper"
-                            Content="Apply GPU Settings" Margin="0,8,0,0"
-                            Click="ButtonGpuApply_Click" />
-                </StackPanel>
-            </Border>
-
             <!-- OTHER SETTINGS -->
             <Border Classes="panel">
                 <StackPanel>
