+++ src/UI/Views/FansWindow.axaml
@@ -13,7 +13,7 @@
 
     <ScrollViewer VerticalScrollBarVisibility="Auto"
                   HorizontalScrollBarVisibility="Disabled">
-    <Grid RowDefinitions="Auto,*,Auto,Auto,Auto">
+    <Grid RowDefinitions="Auto,*,Auto">
 
         <!-- HEADER -->
         <Border Grid.Row="0" Classes="panel-header" Margin="10,10,10,5">
@@ -22,6 +22,8 @@
                     <TextBlock x:Name="headerFanCurves" Classes="header" Text="Fan Curves" FontSize="14" />
                     <TextBlock x:Name="labelMode" Classes="label-dim"
                                Text="Mode: Balanced" />
+                    <TextBlock x:Name="labelSensors" Classes="label-dim"
+                               Text="" Margin="0,4,0,0" />
                 </StackPanel>
                 <Button Grid.Column="1" x:Name="buttonMonitor" Classes="ghelper"
                         Margin="4,0" Height="30"
@@ -73,108 +75,8 @@
         </Grid>
 
         <!-- POWER LIMITS -->
-        <Border Grid.Row="2" Classes="panel" Margin="10,5,10,10">
-            <StackPanel>
-                <TextBlock x:Name="headerPowerLimits" Classes="header" Text="Power Limits" Margin="0,0,0,8" />
-
-                <!-- PPT SPL (PL1) -->
-                <Grid ColumnDefinitions="120,*,50" Margin="0,4">
-                    <TextBlock Grid.Column="0" x:Name="labelPL1Label" Text="CPU PL1 (SPL)"
-                               Classes="label-dim" VerticalAlignment="Center" />
-                    <Slider Grid.Column="1" x:Name="sliderPL1"
-                            Minimum="5" Maximum="120" Value="45"
-                            TickFrequency="5"
-                            ValueChanged="SliderPL1_ValueChanged" />
-                    <TextBlock Grid.Column="2" x:Name="labelPL1" Classes="value"
-                               Text="45W" HorizontalAlignment="Center"
-                               VerticalAlignment="Center" />
-                </Grid>
-
-                <!-- PPT SPPT (PL2) -->
-                <Grid ColumnDefinitions="120,*,50" Margin="0,4">
-                    <TextBlock Grid.Column="0" x:Name="labelPL2Label" Text="CPU PL2 (SPPT)"
-                               Classes="label-dim" VerticalAlignment="Center" />
-                    <Slider Grid.Column="1" x:Name="sliderPL2"
-                            Minimum="5" Maximum="150" Value="65"
-                            TickFrequency="5"
-                            ValueChanged="SliderPL2_ValueChanged" />
-                    <TextBlock Grid.Column="2" x:Name="labelPL2" Classes="value"
-                               Text="65W" HorizontalAlignment="Center"
-                               VerticalAlignment="Center" />
-                </Grid>
-
-                <!-- PPT fPPT (fast boost) -->
-                <Grid x:Name="gridFppt" ColumnDefinitions="120,*,50" Margin="0,4" IsVisible="False">
-                    <TextBlock Grid.Column="0" x:Name="labelFpptLabel" Text="CPU fPPT (Fast)"
-                               Classes="label-dim" VerticalAlignment="Center" />
-                    <Slider Grid.Column="1" x:Name="sliderFppt"
-                            Minimum="5" Maximum="150" Value="65"
-                            TickFrequency="5"
-                            ValueChanged="SliderFppt_ValueChanged" />
-                    <TextBlock Grid.Column="2" x:Name="labelFppt" Classes="value"
-                               Text="65W" HorizontalAlignment="Center"
-                               VerticalAlignment="Center" />
-                </Grid>
-
-                <!-- CPU Boost -->
-                <Grid ColumnDefinitions="120,*" Margin="0,8,0,0">
-                    <TextBlock Grid.Column="0" x:Name="labelCpuBoostLabel" Text="CPU Boost"
-                               Classes="label-dim" VerticalAlignment="Center" />
-                    <StackPanel Grid.Column="1" Orientation="Horizontal" Spacing="4">
-                        <Button x:Name="buttonBoostOff" Classes="ghelper"
-                                Content="Off" Click="ButtonBoostOff_Click" />
-                        <Button x:Name="buttonBoostOn" Classes="ghelper"
-                                Content="On" Click="ButtonBoostOn_Click" />
-                    </StackPanel>
-                </Grid>
-
-                <!-- Auto-apply power -->
-                <CheckBox x:Name="checkApplyPower"
-                          Content="Auto-apply power limits on mode change" Margin="0,8,0,0"
-                          IsCheckedChanged="CheckApplyPower_Changed" />
-
-                <!-- Sensor readings -->
-                <TextBlock x:Name="labelSensors" Classes="label-dim"
-                           Text="" Margin="0,8,0,0" />
-            </StackPanel>
-        </Border>
-
-        <!-- Undervolting (Ryzen CO) -->
-        <Border Grid.Row="3" x:Name="panelUV" Classes="panel" Margin="10,0,10,10" IsVisible="False">
-            <StackPanel>
-                <TextBlock x:Name="headerUndervolt" Classes="header" Text="Undervolting (Experimental)" Margin="0,0,0,4" />
-                <TextBlock x:Name="labelUndervoltDesc" Classes="label-dim" TextWrapping="Wrap" Margin="0,0,0,8"
-                           Text="Curve Optimizer offset. More negative = lower voltage. Too-low values can cause instability or data corruption. Start small. Resets on reboot." />
-
-                <Grid ColumnDefinitions="120,*,50" Margin="0,4">
-                    <TextBlock Grid.Column="0" x:Name="labelUndervoltCpu" Text="CPU"
-                               Classes="label-dim" VerticalAlignment="Center" />
-                    <!-- Slider is positive 0..40 (intensity) so the fill bar grows
-                         left-to-right with UV strength; value is negated when stored. -->
-                    <Slider Grid.Column="1" x:Name="sliderCpuUV"
-                            Minimum="0" Maximum="40" Value="0"
-                            TickFrequency="1" IsSnapToTickEnabled="True"
-                            ValueChanged="SliderCpuUV_ValueChanged" />
-                    <TextBlock Grid.Column="2" x:Name="labelCpuUV" Classes="value"
-                               Text="0" HorizontalAlignment="Center"
-                               VerticalAlignment="Center" />
-                </Grid>
-
-                <StackPanel Orientation="Horizontal" Spacing="8" Margin="0,8,0,0">
-                    <Button x:Name="buttonApplyUV" Classes="ghelper"
-                            Content="Apply" Click="ButtonApplyUV_Click" />
-                    <Button x:Name="buttonResetUV" Classes="ghelper"
-                            Content="Reset" Click="ButtonResetUV_Click" />
-                    <CheckBox x:Name="checkApplyUV"
-                              Content="Auto-apply on mode change"
-                              VerticalAlignment="Center"
-                              IsCheckedChanged="CheckApplyUV_Changed" />
-                </StackPanel>
-            </StackPanel>
-        </Border>
-
         <!-- Advanced: per-mode shell hook + power-limit reapply timer. -->
-        <Border Grid.Row="4" Classes="panel" Margin="10,0,10,10">
+        <Border Grid.Row="2" Classes="panel" Margin="10,0,10,10">
             <StackPanel>
                 <TextBlock x:Name="headerAdvanced" Classes="header" Text="Advanced" Margin="0,0,0,4" />
